Add the ability to attach the logs to the ticket created on bugzilla
For the moment this change only attach the build_log to the ticket, but
simply change the line #136 and it will attach any file you want.

To attach the file, it first downloads them into a temp file that is
then removed automatically.

All files are attached as text/plain to make them viewable directly on
bugzilla.