Fix getting the right logs URLs
Until now we were trying to use the parent task to access the logs while
the parent task has no log, it's just the parent task.
Now we are retrieving the children tasks of the parent, finding which one
of the children failed building (the first one we find) and from there we
are generating the URLs to the log files.