Reduce the number of calls made by update-critpath
The pkgdb2 API is able to handle queries submitting multiple packages at
once, so there is no need to make a query for each package separately.

In addition this commit makes sure there are packages to change before
trying to do the change and will print a message if there are no packages
to change.