run-pungi: atomic: Change the `ostree_osname` to be fedora-atomic
We should be consistent with the cloud image kickstart,
which uses `fedora-atomic`.

Note this string actually appears in the user interface and may be
typed by the admins, so the shift key gymnastics required by the
previous value are more painful than just having it in a URL or the
like.