build-functions: Generate OSTree static deltas for rawhide
Adam Miller pointed out that OSTree downloads are slow.  We've been
working on "static deltas" for a while now, and they're ready for some
real world testing.

For more information, see:

 - https://mail.gnome.org/archives/ostree-list/2015-March/msg00005.html
 - https://bugzilla.gnome.org/show_bug.cgi?id=721799

This is a *basic* implementation - a more sophisticated script is
necessary to generate "spanning deltas" across multiple revisions.

The deltas make more sense for releases that are slower than daily -
for example, there's discussion of doing 2 week stable Fedora Atomic
Host updates.

In that model, it'd be much saner to generate deltas from say the last
update, and possibly the update before it.

For "rolling" repositories, the audience is far smaller, and that's
exactly what the incrementalism of the default OSTree archive-z2
repository is designed for.

The delta size is going to vary based on the input.  Some numbers are
in the bugzilla.  For reference, a static delta based on the the
latest rawhide commit to the previous is 5MB.

Signed-off-by: Colin Walters <walters@verbum.org>