we now hash debug, drpms and SRPMS so find them in the path name
we want to skip over the srpms drpms and debuginfo rpms so skip
over them, the patch used to always end in their name but that is
no longer true.  we could chop off the last number/letter and /
and still check the end but there should never be debug SRPM delta
in the path for any reason other than a repo containing the things
we do not need to check.

Signed-off-by: Dennis Gilmore <dennis@ausil.us>